Hide Connected Accounts section on Account page
