+++ src/applications/personalization/account/reducers/index.js
@@ -2,6 +2,7 @@
 // since that is what the rest of the applications have known these properties as, even though from
 // this point forward these attributes will be known as "account" information since it mostly comes
 // from the IDP.
-import profileInformation from '../../../../platform/user/profile/reducers';
+import profileInformation from 'platform/user/profile/reducers';
+import connectedAccounts from 'applications/personalization/connected-accounts/reducers/connectedAccounts';
 
-export default { account: profileInformation };
+export default { account: profileInformation, connectedAccounts };
